Add files via upload
This undertaking entails the advent of a easy Ethereum clever settlement with simple garage capability and a frontend utility to interact with the agreement. The smart settlement permits users to store and retrieve various and a message. The frontend, constructed with HTML and JavaScript, offers an intuitive interface for interacting with the clever settlement.